brentru fork for esp8266, wsign compare

+++ src/components/ds18x20/ws_ds18x20.cpp
@@ -135,7 +135,7 @@ bool ws_ds18x20::addDS18x20(
 void ws_ds18x20::deleteDS18x20(
     wippersnapper_ds18x20_v1_Ds18x20DeInitRequest *msgDS18x20DeinitReq) {
   // Loop thru vector of drivers to find the unique address
-  for (int idx = 0; idx < _ds18xDrivers.size(); idx++) {
+  for (size_t idx = 0; idx < _ds18xDrivers.size(); idx++) {
     if (strcmp(_ds18xDrivers[idx]->onewire_pin,
                msgDS18x20DeinitReq->onewire_pin) == 0) {
       WS_DEBUG_PRINT("Deleting OneWire instance on pin ");
